Refactored mavlink code into datalogger and mavlink sensor.
